fix: move StampSourcePipelineTypeLabelOnCreate from cascadedelete to go/api
