+++ go/cascadedelete/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
     importpath = "github.com/michelangelo-ai/michelangelo/go/cascadedelete",
     visibility = ["//visibility:public"],
     deps = [
-        "//go/api:go_default_library",
         "@com_github_prometheus_client_golang//prometheus:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
@@ -35,7 +34,6 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//go/api:go_default_library",
         "//proto/api/v2:go_default_library",
         "@com_github_prometheus_client_golang//prometheus:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
