+++ go/cascadedelete/apihook.go
@@ -6,8 +6,6 @@ import (
 	"go.uber.org/zap"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	"github.com/michelangelo-ai/michelangelo/go/api"
 )
 
 // StampOwnerRefOnCreate stamps owner as the controller ownerReference on child
@@ -37,20 +35,4 @@ func StampOwnerRefOnCreate(ctx context.Context, logger *zap.Logger, scheme *runt
 			zap.String("child", child.GetName()))
 	}
 	return nil
-}
-
-// StampSourcePipelineTypeLabelOnCreate stamps pipelineType (e.g.
-// "PIPELINE_TYPE_TRAIN") onto child under
-// api.SourcePipelineTypeLabelName. No-op if pipelineType is empty.
-func StampSourcePipelineTypeLabelOnCreate(child client.Object, pipelineType string) {
-	if pipelineType == "" {
-		return
-	}
-
-	labels := child.GetLabels()
-	if labels == nil {
-		labels = map[string]string{}
-	}
-	labels[api.SourcePipelineTypeLabelName] = pipelineType
-	child.SetLabels(labels)
 }
