+++ go/components/pipelinerun/apihook/apihook.go
@@ -87,7 +87,7 @@ func (a apiHook) BeforeCreate(ctx context.Context, request *v2.CreatePipelineRun
 	}
 
 	if pipelineType := pipeline.Spec.GetType(); pipelineType != v2.PIPELINE_TYPE_INVALID {
-		cascadedelete.StampSourcePipelineTypeLabelOnCreate(request.PipelineRun, pipelineType.String())
+		api.StampSourcePipelineTypeLabelOnCreate(request.PipelineRun, pipelineType.String())
 	}
 
 	return cascadedelete.StampOwnerRefOnCreate(ctx, a.logger, a.scheme, request.PipelineRun, pipeline)
